+++ CHANGELOG.md
@@ -24,6 +24,10 @@
 - (Feature) (Scheduler) Deployment Scale Functionality
 - (Feature) (Platform) Chart Integration
 - (Maintenance) Switch to google.golang.org/protobuf
+- (Feature) Add DebugPackage to the OPS Binary
+- (Feature) (Networking) ArangoRoute Protocol
+- (Feature) (Platform) Platform Requirements support
+- (Improvement) Drop slash requirement from ArangoRoute
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD